Added Files through upload
Added the folder created by scrapy, as well as a start of a spider that crawls the music subreddit. As of right now, the "redMusic" spider scrapes the home page of /r/Music for the titles and vote total, then loads the information scraped into the csv file labelled "redditMusic. Moving forward this will be modified to follow each of the links on this homepage, and scrape all comments.